Store registered operation hashes in persistence state
