+++ Sources/GraphQLCodegen/Input/Documents/Documents.swift
@@ -1,4 +1,3 @@
-import CryptoKit
 import Foundation
 
 struct Documents {
@@ -25,23 +24,15 @@ struct Document: Sendable {
     }
 
     struct Operation: Sendable {
+        enum Persistence: Sendable {
+            case standard
+            case registered(hash: String)
+        }
+
         let ast: AST.OperationDefinition
         let canonicalText: String
+        let persistence: Persistence
         let sourceText: Substring
-
-        var hash: String {
-            let digits = Array("0123456789abcdef".utf8)
-            let capacity = 2 * SHA256.Digest.byteCount
-            return String(unsafeUninitializedCapacity: capacity) { buffer -> Int in
-                var index = 0
-                for byte in SHA256.hash(data: Data(canonicalText.utf8)) {
-                    buffer[index] = digits[Int(byte >> 4)]
-                    buffer[index + 1] = digits[Int(byte & 0x0F)]
-                    index += 2
-                }
-                return capacity
-            }
-        }
     }
 
     struct Fragment {
